requirements.txt
azure-ai-textanalytics>=5.1.0
python-dotenv>=0.21.0
